requirements.txt
# Core Data Science
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0
joblib>=1.3.0

# Explainability
shap>=0.44.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Streamlit App
streamlit>=1.28.0
streamlit-extras>=0.3.0

# Data Download
kagglehub>=0.2.0

# API & Environment
python-dotenv>=1.0.0
requests>=2.31.0

# Utilities
Pillow>=10.0.0
